Fix aramUploadData callback param in the source file

+++ src/musyx/runtime/hw_aramdma.c
@@ -82,7 +82,7 @@ static void aramQueueCallback(unsigned long ptr) {
 }
 
 void aramUploadData(void* mram, unsigned long aram, unsigned long len, unsigned long highPrio,
-                    void (*callback)(unsigned long), unsigned long user) {
+                    void (*callback)(size_t), unsigned long user) {
   ARAMTransferQueue* aramQueue; // r31
   int old;                      // r30
